Add Unit Tests
Well, they are technically integration tests. This change also involved a bit of a refactor of moving the API code out of the Flask Handler in an attempt to support mocks, but that did not end well (darn you Postgres).

On the bright side, we also have MyPy and Flake8 running to make sure things are happy-ish.